Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gongjungjebiyudanja <73692488+YoungHyeon1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
boto3 = "^1.34.118"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
lib-commons = {path = "../lib-commons"}
pandas = "^2.2.2"
pyathena = "^3.8.3"
ipykernel = "^6.29.4"
chompjs = "^1.2.3"
psycopg = {extras = ["binary"], version = "^3.1.19"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
